fix: avoid recursive form preload
